# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-SRO",
    "summary": "Workflow Builder: demonstrations in, reviewable skills out.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/skills/induct": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Induce Skill",
        "description": "Two runs to one skill version. Runs synchronously: induction is pure\ncomputation over sealed recordings and takes milliseconds. The durable path\nthrough Temporal exists for callers that want a retryable handle.",
        "operationId": "induce_skill_v1_skills_induct_post",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InduceSkillRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InductionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List Skills",
        "operationId": "list_skills_v1_skills_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillSummary"
                  },
                  "title": "Response List Skills V1 Skills Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/{skill_id}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Skill",
        "operationId": "get_skill_v1_skills__skill_id__get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/skills/{skill_id}/promote": {
      "post": {
        "tags": [
          "skills"
        ],
        "summary": "Promote Skill",
        "operationId": "promote_skill_v1_skills__skill_id__promote_post",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssertionModel": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "expected": {
            "type": "string",
            "title": "Expected"
          },
          "pointer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pointer"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "expected",
          "pointer"
        ],
        "title": "AssertionModel"
      },
...
      "Direction": {
        "type": "string",
        "enum": [
          "inbound",
          "outbound",
          "internal"
        ],
        "title": "Direction"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InduceSkillRequest": {
        "properties": {
          "first_recording_id": {
            "type": "string",
            "title": "First Recording Id"
          },
          "second_recording_id": {
            "type": "string",
            "title": "Second Recording Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "first_recording_id",
          "second_recording_id"
        ],
        "title": "InduceSkillRequest"
      },
      "InductionResponse": {
        "properties": {
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count"
          },
          "input_parameter_count": {
            "type": "integer",
            "title": "Input Parameter Count"
          },
          "derived_parameter_count": {
            "type": "integer",
            "title": "Derived Parameter Count"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "version",
          "step_count",
          "input_parameter_count",
          "derived_parameter_count"
        ],
        "title": "InductionResponse"
      },
      "NetworkPlanModel": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "expected_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Status"
          },
          "replayable": {
            "type": "boolean",
            "title": "Replayable"
          },
          "unreplayable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unreplayable Reason"
          },
          "required_credentials": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Credentials"
          }
        },
        "type": "object",
        "required": [
          "method",
          "url",
          "headers",
          "body",
          "expected_status",
          "replayable",
          "unreplayable_reason",
          "required_credentials"
        ],
        "title": "NetworkPlanModel"
      },
      "ObjectiveKeyModel": {
        "properties": {
          "objective_type": {
            "type": "string",
            "title": "Objective Type"
          },
          "target_system": {
            "type": "string",
            "title": "Target System"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "facility": {
            "type": "string",
            "title": "Facility"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        },
        "type": "object",
        "required": [
          "objective_type",
          "target_system",
          "entity_type",
          "facility",
          "direction"
        ],
        "title": "ObjectiveKeyModel"
      },
      "ParameterModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "observed_values": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observed Values"
          },
          "source_step_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Step Index"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "description",
          "observed_values",
          "source_step_index"
        ],
        "title": "ParameterModel"
      },
      "PromoteRequest": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "to": {
            "type": "string",
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "version",
          "to"
        ],
        "title": "PromoteRequest"
      },
...
      "SkillDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "objective_key": {
            "$ref": "#/components/schemas/ObjectiveKeyModel"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "latest_version": {
            "type": "integer",
            "title": "Latest Version"
          },
          "latest_stage": {
            "type": "string",
            "title": "Latest Stage"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/SkillVersionModel"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "objective_key",
          "created_at",
          "latest_version",
          "latest_stage",
          "versions"
        ],
        "title": "SkillDetail"
      },
      "SkillSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "objective_key": {
            "$ref": "#/components/schemas/ObjectiveKeyModel"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "latest_version": {
            "type": "integer",
            "title": "Latest Version"
          },
          "latest_stage": {
            "type": "string",
            "title": "Latest Stage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "objective_key",
          "created_at",
          "latest_version",
          "latest_stage"
        ],
        "title": "SkillSummary"
      },
      "SkillVersionModel": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "induced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Induced At"
          },
          "induced_by": {
            "type": "string",
            "title": "Induced By"
          },
          "recording_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recording Ids"
          },
          "provenance_note": {
            "type": "string",
            "title": "Provenance Note"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepModel"
            },
            "type": "array",
            "title": "Steps"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/ParameterModel"
            },
            "type": "array",
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "version",
          "stage",
          "induced_at",
          "induced_by",
          "recording_ids",
          "provenance_note",
          "steps",
          "parameters"
        ],
        "title": "SkillVersionModel"
      },
...
      "StepModel": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "requires_human": {
            "type": "boolean",
            "title": "Requires Human"
          },
          "network_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NetworkPlanModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "ui_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UiPlanModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "assertions": {
            "items": {
              "$ref": "#/components/schemas/AssertionModel"
            },
            "type": "array",
            "title": "Assertions"
          }
        },
        "type": "object",
        "required": [
          "index",
          "intent",
          "requires_human",
          "network_plan",
          "ui_plan",
          "assertions"
        ],
        "title": "StepModel"
      },
      "UiPlanModel": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "target_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Path"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "wait_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait For"
          }
        },
        "type": "object",
        "required": [
          "action",
          "target",
          "target_path",
          "value",
          "wait_for"
        ],
        "title": "UiPlanModel"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}